# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "DocForge",
        "version": "0.1.0"
    },
    "paths": {
...
        "/v1/documents/{document_id}/review": {
            "get": {
                "summary": "Review Detail",
                "operationId": "review_detail_v1_documents__document_id__review_get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Sign",
                "operationId": "sign_v1_documents__document_id__review_post",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SignIn"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SignedOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/documents/{document_id}/claim": {
            "post": {
                "summary": "Claim",
                "description": "Claim the review for a few minutes, or renew the claim; call again to keep it.\n\nWhile another reviewer's claim lasts it is returned unchanged (`mine` false), and\ntheir changes alone are accepted; an administrator may `take_over`.",
                "operationId": "claim_v1_documents__document_id__claim_post",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClaimIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClaimOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Release",
                "description": "End the caller's claim. Someone else's is left alone (still 204).",
                "operationId": "release_v1_documents__document_id__claim_delete",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/documents/{document_id}/corrections": {
            "post": {
                "summary": "Correct",
                "operationId": "correct_v1_documents__document_id__corrections_post",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CorrectionIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewOut"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/documents/{document_id}/pages/{page}": {
            "get": {
                "summary": "Page Image",
                "operationId": "page_image_v1_documents__document_id__pages__page__get",
                "parameters": [
                    {
                        "name": "document_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "title": "Document Id"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Page"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "image/png": {}
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClaimIn": {
                "properties": {
                    "take_over": {
                        "type": "boolean",
                        "title": "Take Over",
                        "default": false
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "title": "ClaimIn"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ReviewOut": {
                "properties": {
                    "document_id": {
                        "type": "string",
                        "format": "uuid",
                        "title": "Document Id"
                    },
                    "doc_type": {
                        "type": "string",
                        "title": "Doc Type"
                    },
                    "filename": {
                        "type": "string",
                        "title": "Filename"
                    },
                    "version_no": {
                        "type": "integer",
                        "title": "Version No"
                    },
                    "page_count": {
                        "type": "integer",
                        "title": "Page Count"
                    },
                    "pages": {
                        "items": {
                            "$ref": "#/components/schemas/docforge__api__review__PageOut"
                        },
                        "type": "array",
                        "title": "Pages"
                    },
                    "decision": {
                        "type": "string",
                        "title": "Decision"
                    },
                    "blockers": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Blockers"
                    },
                    "record": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Record"
                    },
                    "assessment": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Assessment"
                    },
                    "editable_paths": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Editable Paths"
                    },
                    "corrections": {
                        "items": {
                            "$ref": "#/components/schemas/CorrectionOut"
                        },
                        "type": "array",
                        "title": "Corrections"
                    },
                    "match_status": {
                        "type": "string",
                        "title": "Match Status"
                    },
                    "discrepancies": {
                        "items": {
                            "additionalProperties": true,
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Discrepancies"
                    },
                    "counterpart_document_id": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "uuid"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Counterpart Document Id"
                    },
                    "review": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/SignedOut"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "signature_valid": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Signature Valid"
                    },
                    "record_sha256": {
                        "type": "string",
                        "title": "Record Sha256"
                    },
                    "meanings": {
                        "additionalProperties": {
                            "type": "string"
                        },
                        "type": "object",
                        "title": "Meanings"
                    },
                    "superseded": {
                        "type": "boolean",
                        "title": "Superseded"
                    },
                    "certificates": {
                        "items": {
                            "additionalProperties": {
                                "type": "string"
                            },
                            "type": "object"
                        },
                        "type": "array",
                        "title": "Certificates"
                    },
                    "claim": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ClaimOut"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "document_id",
                    "doc_type",
                    "filename",
                    "version_no",
                    "page_count",
                    "pages",
                    "decision",
                    "blockers",
                    "record",
                    "assessment",
                    "editable_paths",
                    "corrections",
                    "match_status",
                    "discrepancies",
                    "counterpart_document_id",
                    "review",
                    "signature_valid",
                    "record_sha256",
                    "meanings",
                    "superseded",
                    "certificates",
                    "claim"
                ],
                "title": "ReviewOut"
            },
...
            "SignIn": {
                "properties": {
                    "outcome": {
                        "type": "string",
                        "enum": [
                            "approved",
                            "rejected"
                        ],
                        "title": "Outcome"
                    },
                    "meaning": {
                        "type": "string",
                        "maxLength": 500,
                        "minLength": 1,
                        "title": "Meaning"
                    },
                    "reason": {
                        "type": "string",
                        "maxLength": 2000,
                        "minLength": 1,
                        "title": "Reason"
                    },
                    "override_reason": {
                        "anyOf": [
                            {
                                "type": "string",
                                "maxLength": 2000
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Override Reason"
                    },
                    "expected_record_sha256": {
                        "type": "string",
                        "pattern": "^[0-9a-f]{64}$",
                        "title": "Expected Record Sha256"
                    },
                    "email": {
                        "type": "string",
                        "maxLength": 320,
                        "title": "Email"
                    },
                    "pin": {
                        "type": "string",
                        "maxLength": 64,
                        "title": "Pin"
                    }
                },
                "additionalProperties": false,
                "type": "object",
                "required": [
                    "outcome",
                    "meaning",
                    "reason",
                    "expected_record_sha256",
                    "email",
                    "pin"
                ],
                "title": "SignIn"
            },
...
        }
    }
}